提取出instantiateComponent模块与强化renderComponent方法
