+++ version.md
@@ -2,7 +2,7 @@
 1. 修正 onChange 事件
 2. 重构 diffProps 模块的实现
 3. 支持组件的isMounted方法
-4. 添加beforeRender, beforePatch, afterPatch钩子
+4. 添加beforePatch , afterPatch钩子
 5. 添加lib/ReactInputSelection.js
 6.  统一所有操作虚拟DOM的方法的参数(mountXXX, updateXXX, alignXXX系列)
 >1 第一个参数为旧真实DOM或旧虚拟DOM
@@ -16,6 +16,8 @@
 10. 完善createStringRef方法，应该能抛错与删除无用数据
 11. 上线全新的任务调度系统
 12. 重构unmountComponentAtNode方法
+13. 添加对两个虚拟DOM的引用都相同的情况下，检测子组件的contextType决定是否更新的策略
+14. 无状态组件支持模块模式（返回一个带生命周期钩子的纯对象，这些方法会像有状态组件那样被调用）
 
 ## 1.1.1
 1. 简化createClass
